git-svn-id: http://svn.airdcpp.net/airsvn/trunk@805 ec3e54a6-43b8-4a3e-9c49-2a733a73d9fe

+++ AutoSearchManager.h
@@ -167,8 +167,8 @@ private:
 	mutable SharedMutex cs;
 
 	//count minutes to be more accurate than comparing ticks every minute.
-	size_t lastSearch;
-	size_t recheckTime;
+	int lastSearch;
+	int recheckTime;
 
 	unsigned int curPos;
